Divide make.js file to small cli files
