+++ .eslintrc.js
@@ -1,0 +1,14 @@
+module.exports = {
+    "env": {
+        "es6": true,
+        "node": true,
+    },
+    "extends": "eslint:recommended",
+    "parserOptions": {
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "rules": {
+        "no-undef": "error"
+    }
+}
