+++ package.json
@@ -33,6 +33,7 @@
   "devDependencies": {
     "adm-zip": "0.4.13",
     "azure-devops-node-api": "^10.2.2",
+    "eslint": "^8.54.0",
     "js-yaml": "^3.13.1",
     "minimatch": "3.0.2",
     "minimist": "^1.2.6",
